add config wrappers
